+++ tests/flow/test_dvid.py
@@ -1,0 +1,3 @@
+
+
+def test_cutout_label():