+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ChangeLog history
 - a plugin system with a median filter example
 - new operator: normalize-intensity
 - support cutout from hdf5 file
+- new operator: cutout-dvid-label
 
 ## Bug Fixes 
 - cutout whole volume in default